+++ mirage/calc/ckd_tree.pxd
@@ -10,7 +10,7 @@ cdef extern from "ckd_tree.h":
       void PointsInCircle(double*, unsigned long, double, double*)
       double MagnificationCoefficient(double, double, double)
       void MagnificationCoefficient(double*, unsigned long, double, double*)
-      vector[int] LensPlaneCoordinates(double, double, double)
+      vector[long] LensPlaneCoordinates(double, double, double)
       unsigned long size()
       unsigned long buf_size()
       unsigned long tree_size()
